TMTNNFR-130: moving utilities to a shared location
