+++ hwprovisioner/resourcemanager/plugins/_utils/filesystem.py
