+++ hwprovisioner/resourcemanager/plugins/_utils/pci.py
@@ -2,7 +2,7 @@ import subprocess
 import re 
 import logging
 import munch
-from infra.utils import shell
+from plugins._utils import shell
 from builtins import staticmethod
 import os
