+++ hwprovisioner/resourcemanager/plugins/_utils/shell.py
