+++ hwprovisioner/resourcemanager/plugins/hypervisor/hypervisor
@@ -1,4 +1,3 @@
 #!/bin/bash
 
-CURRENT_DIR=$(dirname ${BASH_SOURCE[0]})
-PYTHONPATH=$PYTHONPATH:$CURRENT_DIR python3 ${CURRENT_DIR}/lab/hypervisor.py $*
+PYTHONPATH=$PYTHONPATH:../../ python3 ./lab/hypervisor.py $*
