+++ hwprovisioner/resourcemanager/plugins/hypervisor/lab/vms/allocator.py
