+++ hwprovisioner/resourcemanager/plugins/hypervisor/lab/vms/test/allocator_test.py
@@ -1,7 +1,7 @@
 import pytest
 from lab.vms import allocator, cloud_init
 from lab.vms import vm_manager, dhcp_handlers
-from infra.utils import pci
+from plugins._utils import pci
 from lab.vms import libvirt_wrapper
 from lab.vms import image_store
 from lab import NotEnoughResourceException
